requirements.txt
attrs
beautifulsoup4
elasticsearch
flake8
click
flask
pytest
mock